+++ components/resource/scenemanager.cpp
@@ -141,9 +141,9 @@ namespace Resource
     class SharedStateManager : public osgDB::SharedStateManager
     {
     public:
-        unsigned int getNumSharedTextures() const { return _sharedTextureList.size(); }
+        size_t getNumSharedTextures() const { return _sharedTextureList.size(); }
 
-        unsigned int getNumSharedStateSets() const { return _sharedStateSetList.size(); }
+        size_t getNumSharedStateSets() const { return _sharedStateSetList.size(); }
 
         void clearCache()
         {
@@ -158,7 +158,7 @@ namespace Resource
     {
     public:
         SetFilterSettingsControllerVisitor(
-            osg::Texture::FilterMode minFilter, osg::Texture::FilterMode magFilter, int maxAnisotropy)
+            osg::Texture::FilterMode minFilter, osg::Texture::FilterMode magFilter, float maxAnisotropy)
             : mMinFilter(minFilter)
             , mMagFilter(magFilter)
             , mMaxAnisotropy(maxAnisotropy)
@@ -170,7 +170,7 @@ namespace Resource
             if (NifOsg::FlipController* flipctrl = dynamic_cast<NifOsg::FlipController*>(&ctrl))
             {
                 for (std::vector<osg::ref_ptr<osg::Texture2D>>::iterator it = flipctrl->getTextures().begin();
-                     it != flipctrl->getTextures().end(); ++it)
+                    it != flipctrl->getTextures().end(); ++it)
                 {
                     osg::Texture* tex = *it;
                     tex->setFilter(osg::Texture::MIN_FILTER, mMinFilter);
@@ -183,7 +183,7 @@ namespace Resource
     private:
         osg::Texture::FilterMode mMinFilter;
         osg::Texture::FilterMode mMagFilter;
-        int mMaxAnisotropy;
+        float mMaxAnisotropy;
     };
 
     /// Set texture filtering settings on textures contained in StateSets.
@@ -191,7 +191,7 @@ namespace Resource
     {
     public:
         SetFilterSettingsVisitor(
-            osg::Texture::FilterMode minFilter, osg::Texture::FilterMode magFilter, int maxAnisotropy)
+            osg::Texture::FilterMode minFilter, osg::Texture::FilterMode magFilter, float maxAnisotropy)
             : osg::NodeVisitor(TRAVERSE_ALL_CHILDREN)
             , mMinFilter(minFilter)
             , mMagFilter(magFilter)
@@ -233,7 +233,7 @@ namespace Resource
     private:
         osg::Texture::FilterMode mMinFilter;
         osg::Texture::FilterMode mMagFilter;
-        int mMaxAnisotropy;
+        float mMaxAnisotropy;
     };
 
     // Check Collada extra descriptions
@@ -461,7 +461,7 @@ namespace Resource
         , mBgsmFileManager(bgsmFileManager)
         , mMinFilter(osg::Texture::LINEAR_MIPMAP_LINEAR)
         , mMagFilter(osg::Texture::LINEAR)
-        , mMaxAnisotropy(1)
+        , mMaxAnisotropy(1.f)
         , mUnRefImageDataAfterApply(false)
         , mParticleSystemMask(~0u)
     {
@@ -1129,7 +1129,7 @@ namespace Resource
     }
 
     void SceneManager::setFilterSettings(
-        const std::string& magfilter, const std::string& minfilter, const std::string& mipmap, int maxAnisotropy)
+        const std::string& magfilter, const std::string& minfilter, const std::string& mipmap, float maxAnisotropy)
     {
         osg::Texture::FilterMode min = osg::Texture::LINEAR;
         osg::Texture::FilterMode mag = osg::Texture::LINEAR;
@@ -1163,7 +1163,7 @@ namespace Resource
 
         mMinFilter = min;
         mMagFilter = mag;
-        mMaxAnisotropy = std::max(1, maxAnisotropy);
+        mMaxAnisotropy = std::max(1.f, maxAnisotropy);
 
         SetFilterSettingsControllerVisitor setFilterSettingsControllerVisitor(mMinFilter, mMagFilter, mMaxAnisotropy);
         SetFilterSettingsVisitor setFilterSettingsVisitor(mMinFilter, mMagFilter, mMaxAnisotropy);
@@ -1225,13 +1225,16 @@ namespace Resource
         if (mIncrementalCompileOperation)
         {
             std::lock_guard<OpenThreads::Mutex> lock(*mIncrementalCompileOperation->getToCompiledMutex());
-            stats->setAttribute(frameNumber, "Compiling", mIncrementalCompileOperation->getToCompile().size());
+            stats->setAttribute(
+                frameNumber, "Compiling", static_cast<double>(mIncrementalCompileOperation->getToCompile().size()));
         }
 
         {
             std::lock_guard<std::mutex> lock(mSharedStateMutex);
-            stats->setAttribute(frameNumber, "Texture", mSharedStateManager->getNumSharedTextures());
-            stats->setAttribute(frameNumber, "StateSet", mSharedStateManager->getNumSharedStateSets());
+            stats->setAttribute(
+                frameNumber, "Texture", static_cast<double>(mSharedStateManager->getNumSharedTextures()));
+            stats->setAttribute(
+                frameNumber, "StateSet", static_cast<double>(mSharedStateManager->getNumSharedStateSets()));
         }
 
         Resource::reportStats("Node", frameNumber, mCache->getStats(), *stats);
