+++ components/stereo/stereomanager.hpp
@@ -78,13 +78,13 @@ namespace Stereo
         //! @Param enableMultiview whether or not to make use of the GL_OVR_Multiview extension, if supported.
         //! @Param near defines distance to near camera clipping plane from view point.
         //! @Param far defines distance to far camera clipping plane from view point.
-        explicit Manager(osgViewer::Viewer* viewer, bool enableStereo, double near, double far);
+        explicit Manager(osgViewer::Viewer* viewer, bool enableStereo, float near, float far);
         ~Manager();
 
         //! Called during update traversal
         void update();
 
-        void updateSettings(double near, double far)
+        void updateSettings(float near, float far)
         {
             mNear = near;
             mFar = far;
@@ -146,8 +146,8 @@ namespace Stereo
         std::shared_ptr<MultiviewFramebuffer> mMultiviewFramebuffer;
         bool mEyeResolutionOverriden;
         osg::Vec2i mEyeResolutionOverride;
-        double mNear;
-        double mFar;
+        float mNear;
+        float mFar;
 
         std::array<View, 2> mView;
         std::array<osg::Matrixd, 2> mViewOffsetMatrix;
